solved conflicts with develop branch
monitoring.html - added missing </thead> error
class input-group-add used in monitoring.html adds a padding of 41px when displaying the url of the API.
This anomaly is only under monitoring tab.
other places (e.g. Settings > Network) the style appears as it should be (20px padding).
This needs to be fixed after Next is merged with Develop